1. Rate Limiting 2. Caching 3. Other fixes

+++ NoviCurrencyWallet.Core/Contracts/IGenericRepository.cs
@@ -3,6 +3,6 @@
 public interface IGenericRepository<T> where T : class
 {
 	public Task<T> AddAsync(T entity);
-	public Task<TResult> GetAsync<TResult>(int? id);
+	public Task<TResult> GetAsync<TResult>(long? id);
 	public Task UpdateAsync(T entity);
 }
