+++ NoviCurrencyWallet.Core/Contracts/IWalletsRepository.cs
@@ -5,7 +5,7 @@ namespace NoviCurrencyWallet.Core.Contracts;
 
 public interface IWalletsRepository : IGenericRepository<Wallet>
 {
-	public Task<GetWalletBalanceDto> GetAsync(int id, string targetCurrency);
+	public Task<GetWalletBalanceDto> GetAsync(long id, string targetCurrency);
 	public Task<GetWalletBalanceDto> CreateWalletAsync(CreateWalletDto createWalletDto);
 
 	public Task AdjustBalance(UpdateWalletBalanceDto updateWalletBalanceDto);
